feat: reorganized dir structure
